requirements.txt
# LangChain Calculator Agent Requirements

# Core LangChain - using specific version that works
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15

# LangChain Hub for prompts
langchainhub>=0.1.15

# OpenAI package for OpenRouter API
openai>=1.10.0

# LangChain OpenAI integration
langchain-openai>=0.1.0

# HTTP client
requests>=2.31.0

# Required dependencies
pydantic>=2.0.0
pydantic-core>=2.0.0

# Optional but recommended for better output
colorama>=0.4.6
python-dotenv>=1.0.0